Always inherit spec from parent device if forceFormatted exists
When a disk perform a PhasePartitioned transition, it checks whether
the child partition is in unprovisioned phase to determine the spec
inheritance should be performed or not. However, the partition cmd
`parted` NDM using also formats the partition with ext4 format at once,
so the PhasePartitioned transition will see the child partition in
PhaseFormatted and does not perform spec inheritance.

This commit fixes by checking the existence of `forceFormatted` to
ensure disks with forceFormatted should pass settings to children.

Signed-off-by: Weihang Lo <weihang.lo@suse.com>