feat: add Longhorn V2 provisioner
This allows adding extra disks for use by Longhorn's V2 data engine.
I've implemented this as a separate LonghornV2Provisioner for the
sake of neatness, rather than grafting it into LonghornV1Provisioner
with a bunch of `if` statements, but some of the implementation
(the UnProvision() and Update() functions) still call back to the
LonghornV1Provisioner to avoid excessive code duplication.

The current implementation theoretically supports using virtio disks,
except that NDM as a whole ignores those, as they don't have WWNs.
This is something we need to address separately as part of
https://github.com/harvester/harvester/issues/6034

Related issue: https://github.com/harvester/harvester/issues/5274

Signed-off-by: Tim Serong <tserong@suse.com>