fix(utils): log mkfs.ext4 error output
Prior to this, if `mkfs.ext4` failed, all we saw was "failed to
format /dev/sda. err: exit status 1".  With this change, we'll
see the command output, which is a bit more useful, for exmaple:
"failed to format /dev/sda. exit status 1: mke2fs 1.46.4
(18-Aug-2021) /dev/sda is apparently in use by the system;
will not make a filesystem here!"

Signed-off-by: Tim Serong <tserong@suse.com>