fix: only sync LHv2 DiskDriver if changed
Previously, every call to LonghornV2Provisioner.Update() would
trigger an udpate of the Longhorn node object and a requeue
even if nothing had changed.  This resulted in the BD getting
suck with the AddedToNode condition set to False and an error:

  admission webhook "validator.longhorn.io" denied the request:
  spec and status of disks on node harvester-node-0 are being
  syncing and please retry later.

Now we only try to sync if the DiskDriver has actually changed,
and everything behaves correctly.

Signed-off-by: Tim Serong <tserong@suse.com>