fix: exclude /dev/dm-* and LHv2 /dev/nvme* devices
When Longhorn V2 volumes are created and attached to VMs, they
appear on Harvester hosts as /dev/dm-* and /dev/nvme* devices.
This is problematic, because NDM thinks those things are actual
disks and creates BD CRs from them.

This change is a bit of a hack until I finish the work for
https://github.com/harvester/harvester/issues/5059.

Signed-off-by: Tim Serong <tserong@suse.com>