fix: set LHv2 DiskDriver to "auto" if unset
We need to force DiskDriver to "auto" if it's not explicitly set,
because Longhorn also does that internally.  If we don't do this,
the subsequent reflect.DeepEqual() in LonghornV2Provisioner.Update()
will always fail because we have an empty string, but the LHN CR
will have it set to "auto" which results in a weird resync loop.

Related issue: https://github.com/harvester/harvester/issues/6709

Signed-off-by: Tim Serong <tserong@suse.com>