Fix LVM provisioner panic when Spec.Devices map is nil (#201)
* Fix LVM provisioner panic when adding device to existing VG with nil Devices map

When an existing LVMVolumeGroup has a nil Spec.Devices map and a new
device is added, the provisioner panics with 'assignment to entry in nil map'.

This fix initializes the Devices map if it's nil before attempting to
add new devices, preventing the panic.

Fixes: panic in addDevOrCreateLVMVgCRD when Spec.Devices is nil
Signed-off-by: jjsair0412 <jjsair0412@naver.com>

* refactor: remove redundant nil check for Devices map in LVMVolumeGroup

The nil check and map initialization for Spec.Devices were redundant
because lvmVG.Spec.Devices is already set earlier and DeepCopy()
copies it correctly to lvmVGCpy. This change simplifies the update logic.

Signed-off-by: jjsair0412 <jjsair0412@naver.com>

---------

Signed-off-by: jjsair0412 <jjsair0412@naver.com>