Validate disk removals
Rejecting disk removals when there are attached volumes in Degraded
state on single node harvester set ups.

We also make sure to reject removals when virtualmachineimages are
in failed state as well.

The logic is split in two parts:
* We take the disk selectors of volumes in degraded state and
compare the selectors against the disktags
* We take the disk selectors of virtualmachineimages in failed
state and compare against disk tags of disk

In both scenarios in case there is match while disk is being
disabled, we list to the user which volumes and then images
needs to get fixed.

Signed-off-by: Martin Dekov <martin.dekov@suse.com>