Validate disk removals
Rejecting disk removals when there are attached volumes in Degraded
state on single node harvester set ups.

We also make sure to reject removals when virtualmachineimages are
in failed state as well.

The logic is split in two parts:
* We take the disk selectors of volumes in degraded state and
compare the selectors against the disktags
* We take the disk selectors of virtualmachineimages in failed
state and compare against disk tags of disk

In both scenarios in case there is match while disk is being
disabled, we list to the user which volumes and then images
needs to get fixed.

Signed-off-by: Martin Dekov <martin.dekov@suse.com>

Address feedback from Moritz

Addressing feedback form Moritz with suggestion to use
get for specific volumes which are degraded.

Initial assumption was that get would reach out kubernetes
API for cache, but he clairified that cache is local so calls
don't weight on the API.

Switched to that.

Signed-off-by: Martin Dekov <martin.dekov@suse.com>

Skip checks in case of missing tags

Skipping checks in case of missing tags as if storage is missing tags
then its either the default one which comes with installation or is
not being used as the tag is the selector which matches the objects
to the disk

Signed-off-by: Martin Dekov <martin.dekov@suse.com>

Fix nil pointer and rbac with node access

Fixing nil pointer by checking the Provisioner of both
old and new objects as those can be nil and then panic
would be handled silently.

Also since we have cache for nodes added added nodes
to the cluster role

Signed-off-by: Martin Dekov <martin.dekov@suse.com>

Add feedback from Jian

Adding feedback from Jian and fixing potential nil pointer
on pv spec CSI, which can be nil before checking it's
selector

Signed-off-by: Martin Dekov <martin.dekov@suse.com>

Address feedback for multiple objects per tag

Addressing additional feedback where I returned the last
problematic object for the tag as opposed to a list of
objects related with the tag.

Also small optimization to skip logic in case vm image
is not in a failed state.

Signed-off-by: Martin Dekov <martin.dekov@suse.com>