Replace virtualmachineimage with backingimage
Replacing validation of blockdevice deletion from virtualmachineimage to
backingimage. Logic rejects disk removals in case there are
backingimages attached in non ready state to the blockdevice. We index
the caches for both longhorn nodes and backingimages to reach all
backingimages related with blockdevices through the longhorn nodes.

Additonal changes
* Replaces fakes and testing for the new objects
* Fixed some RBAC permissions for the new objects
* Registered longhorn factory as caches didn't actually return
anything
* Removed unsued objects and prints
* Fixed selection based on Spec.Tags to Status.Tags as in real
environment tags are stored in status, not in the spec

Signed-off-by: Martin Dekov <martin.dekov@suse.com>