Validate disk removals
Validating disk removals in case the disk contains
the volume with last healthy replica. The healthy
replica is determined by the FailedAt. In case this
field is empty, then the replica is deemed healthy.

Also rejection happens in case it hosts the last
usable backing image. "Usable backing image"
is one which in ready state.

Signed-off-by: Martin Dekov <martin.dekov@suse.com>