Rename pipeline integration test step
Renaming pipeline integration image build step from:
"Build the Image for the Integration Test"
to:
"Build Integration Test Image and run Unit Tests"

To make it obvious when unit tests are being ran as
we don't mention this anywhere explicitly. Due to this
also removed the `unit-test` script and step.

Signed-off-by: Martin Dekov <martin.dekov@suse.com>