Validate disk removals (#248)
* Validate disk removals

Validating disk removals in case the disk contains the
volume with last healthy replica. The healthy replica is
determined by the FailedAt and HealthyAt. In case
FailedAt field is empty and HealthyAt field is not empty,
then the replica is deemed healthy following examples
directly from longhorn project.

Also rejection happens in case it hosts the last
usable backing image. "Usable backing image"
is one which is in ready state.

Signed-off-by: Martin Dekov <martin.dekov@suse.com>